+++ .gitignore
@@ -1,3 +1,32 @@
-*.pyc
+# Environments
+.env
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+
+# LSP config files
+pyrightconfig.json
+
+### VisualStudioCode ###
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+!.vscode/*.code-snippets
+
+# Git
 .git/
+
+# Python Cache
+*.pyc
 __pycache__/
+
+# Input/Output data
+*.bin
+*.png
+*.jpeg
+*.jpg
