+++ run.py
@@ -10,98 +10,6 @@ from scipy.ndimage import gaussian_filter
 import time
 import numpy as np
 
-class Parameters:
-    def __init__(self, parameters: list):
-        self.model_id = parameters[0]
-        self.export_model_to_binary_file = parameters[1]
-        self.plot_model_bool = parameters[2]
-        self.model_smoothing_bool = parameters[3]
-        self.smooth_level = parameters[4]
-        self.binary_model_path = parameters[5]
-        self.vs_velocity_approximation = parameters[6]
-        self.rho_value_approximation = parameters[7]
-        self.image_to_model_id = parameters[8]
-        self.model_extra_routine = parameters[9]
-        self.image_file_path = parameters[10]
-        self.vp_velocity = parameters[11]
-        self.vs_velocity = [round(i / 1.7, 2) for i in parameters[11]] if parameters[6] else parameters[12]
-        self.rho_value = [round((0.31 * i ** 0.25) * 1e3, 2) for i in parameters[11]] if parameters[7] else parameters[13]
-        self.complex_model_bool = parameters[14]
-        self.inverse_velocity = parameters[15]
-        self.vpmin = parameters[16]
-        self.vpmax = parameters[17]
-        self.vsmin = parameters[18] if not parameters[6] else round(parameters[16] / 1.7, 2)
-        self.vsmax = parameters[19] if not parameters[6] else round(parameters[17] / 1.7, 2)
-        self.rhomin = parameters[20] if not parameters[7] else round((0.31 * parameters[16] ** 0.25) * 1e3, 2)
-        self.rhomax = parameters[21] if not parameters[7] else round((0.31 * parameters[17] ** 0.25) * 1e3, 2)
-        self.parallel_plane_model_id = parameters[22]
-        self.nx = parameters[23]
-        self.nz = parameters[24]
-        self.interfaces = parameters[25]
-        self.vp_interfaces = parameters[26]
-        self.vs_interfaces = parameters[27]
-        self.rho_interfaces = parameters[28]
-
-def print_parameters(data: Parameters) -> None:
-    print("=============== Configuration ===============\n")
-    
-    print("### General Parameters ###\n")
-    print("-> Model Parameters")
-    print(f"        Model_ID = {data.model_id}")
-    print(f"        Export_Model_to_Binary_File = {data.export_model_to_binary_file}")
-    print(f"        Plot_Model = {data.plot_model_bool}")
-    print(f"        Model_Smoothing = {data.model_smoothing_bool}")
-    print(f"        Smooth_Level = {data.smooth_level}\n")
-    
-    print("-> Export Model Path")
-    print(f"        Binary_Model_Path = {data.binary_model_path}\n")
-    
-    print("-> VS Velocity Approximation")
-    print(f"        VS_Velocity = {data.vs_velocity_approximation}\n")
-    
-    print("-> Density Value Approximation")
-    print(f"        Density_Value = {data.rho_value_approximation}\n")
-
-    print("### Image to Model Area ###\n")
-    print("-> Model Bools")
-    print(f"        Image_To_Model_ID = {data.image_to_model_id}")
-    print(f"        Model_Extra_Routine = {data.model_extra_routine}\n")
-    
-    print("-> Load Image Model")
-    print(f"        Image_File_Path = {data.image_file_path}\n")
-    
-    print("-> Model Parameters")
-    print(f"        VP_Velocity = {data.vp_velocity}")
-    print(f"        VS_Velocity = {data.vs_velocity}")
-    print(f"        Rho_Value = {data.rho_value}\n")
-    
-    print("### Complex Image Model ###\n")
-    print("-> Model Bools")
-    print(f"        Complex_Model_ID = {data.complex_model_bool}")
-    print(f"        Inverse_Velocity = {data.inverse_velocity}\n")
-
-    print("-> Model Parameters")
-    print(f"        Minimum_VP_Velocity = {data.vpmin}")
-    print(f"        Maximum_VP_Velocity = {data.vpmax}")
-    print(f"        Minimum_VS_Velocity = {data.vsmin}")
-    print(f"        Maximum_VS_Velocity = {data.vsmax}")
-    print(f"        Minimum_Density = {data.rhomin}")
-    print(f"        Maximum_Density = {data.rhomax}\n")
-    
-    print("### Parallel Plane Model Area ###\n")
-    print("-> Model Bools")
-    print(f"        Parallel_Plane_Model_ID = {data.parallel_plane_model_id}\n")
-    
-    print("-> Model Parameters")
-    print(f"        Nx = {data.nx}")
-    print(f"        Nz = {data.nz}")
-    print(f"        Interfaces = {data.interfaces}")
-    print(f"        VP_Velocity = {data.vp_interfaces}")
-    print(f"        VS_Velocity = {data.vs_interfaces}")
-    print(f"        Rho_Value = {data.rho_interfaces}\n")
-    
-    print("============================================\n")
-
 def model_type_img(id: int, image: np.array, frequent_brightness: list, vp_interfaces: list, vs_interfaces: list, rho_interfaces: list, model_routine_bool: bool) -> np.array:
     if id == 1:
         return ModelCreator.create_model(image, frequent_brightness, vp_interfaces, model_routine_bool)
