+++ src/create/img2model/model.py
@@ -1,3 +1,5 @@
+from src import np, ImageDimensions, ModelRoutine
+
 class CreateModel(ImageDimensions):
     """
     Image -> Model conversion 
@@ -10,7 +12,7 @@ class CreateModel(ImageDimensions):
         self.model = np.zeros((self.height, self.width))
         self.routine_bool = routine_bool
 
-    def set_model_values(self) -> np.array:
+    def set_values(self) -> np.array:
         for i in range(self.height):
             for j in range(self.width):
                 brightness = self.image[i][j]
@@ -32,24 +34,3 @@ class CreateModel(ImageDimensions):
         """
         return np.argmin([abs(element - brightness.astype(np.uint16)) for element in frequent_brightness])
 
-class CreateComplexModel(ImageDimensions):
-    def __init__(self, image: np.array, pmin: float, pmax: float, inverse_velocity: bool) -> None:
-        super().__init__(image)
-        self.pmin = pmin
-        self.pmax = pmax
-        self.inverse_velocity = inverse_velocity
-        self.model = np.zeros((self.height, self.width))
-
-    def set_model_values(self):
-        min_brightness = np.min(self.image)
-        max_brightness = np.max(self.image)
-        v_ratio = (self.pmax - self.pmin) / (max_brightness - min_brightness)
-        for i in range(self.height):
-            for j in range(self.width):
-                brightness = self.image[i][j]
-                self.model[i][j] = self.__get_velocity_order(brightness, v_ratio, min_brightness)
-        return self.model
-
-    def __get_velocity_order(self, brightness, v_ratio, min_brightness):
-        arg = (brightness - min_brightness) * v_ratio
-        return self.pmax - arg if self.inverse_velocity else self.pmin + arg
