+++ utils/__init__.py
@@ -1,0 +1,5 @@
+from .error_handling import ErrorHandling
+
+__all__ = [
+    "ErrorHandling"
+]
