+++ utils/error_handling.py
@@ -2,31 +2,26 @@ class ErrorHandling:
     """
     Class for management of error messages
     """
+
     @staticmethod
     def verify_image_existence() -> None:
-        print("Image not found. Please verify the file path.")
-        exit()
+        raise FileNotFoundError("Image not found. Please verify the file path.")
 
     @staticmethod
     def wrong_image_format() -> None:
-        print("Please use a valid image format. [png, jpg, jpeg]")
-        exit()
+        raise ValueError("Please use a valid image format. [png, jpg, jpeg]")
 
     @staticmethod
     def could_not_open_file() -> None:
-        print("Can't open/read file. Make sure your file path is right.")
-        exit()
-# error handling parallel
-# class ErrorHandling:
-#     """
-#     Class for management of error messages
-#     """
-#     @staticmethod
-#     def interface_error(interfaces: list, value_interfaces: list) -> None:
-#         if len(interfaces) != len(value_interfaces) - 1:
-#             raise ValueError("Interfaces Must be a Length Smaller than velocity_interfaces!")
-#
-#     @staticmethod
-#     def negative_model_parameters(Nx: int, Nz: int) -> None:
-#         if Nx < 0 or Nz < 0:
-#             raise ValueError("Model Parameters Cannot be Negative!")
+        raise IOError("Can't open/read file. Make sure your file path is right.")
+
+    @staticmethod
+    def interface_error(interfaces: list, value_interfaces: list) -> None:
+        if len(interfaces) != len(value_interfaces) - 1:
+            raise ValueError("Interfaces must be a length smaller than velocity_interfaces!")
+
+    @staticmethod
+    def negative_model_parameters(Nx: int, Nz: int) -> None:
+        if Nx < 0 or Nz < 0:
+            raise ValueError("Model parameters cannot be negative!")
+
